fix bug

+++ app/src/main/java/com/isen/math_hunt/activities/GeoGroupActivity.java
@@ -145,7 +145,7 @@ public class GeoGroupActivity extends AppCompatActivity implements LocationListe
 
 
         int dist = (int) distance(geoGroupPosX, location.getLatitude(), geoGroupPosY, location.getLongitude());
-        if (dist > geoGroupRadius) { // changer la valeur par geoGroupRadius
+        if (dist > geoGroupRadius.intValue()) { // changer la valeur par geoGroupRadius
             geoGroupContinueButton.setEnabled(false);
             geoGroupContinueButton.setText("Encore un peu de marche!");
